Cleaned up log reader and memory service

+++ Simple.Wpf.Composition/Services/IMemoryService.cs
@@ -4,7 +4,7 @@
 
     public interface IMemoryService
     {
-        IObservable<long> MemoryInBytes { get;  }
+        IObservable<decimal> MemoryInBytes { get; }
 
         IObservable<decimal> MemoryInKiloBytes { get;  }
